+++ include/pros/tmei.h
@@ -19,22 +19,18 @@
 #ifndef _PROS_TMEI_H_
 #define _PROS_TMEI_H_
 
-#include <stdarg.h>  // va_list, va_start, va_end
-#include <stdbool.h> // bool
+#include <stdarg.h>   // va_list, va_start, va_end
+#include <stdbool.h>  // bool
 #define _GNU_SOURCE
 // NOTE: this would normally be in the C file, but it won't compile that way
-#include <stdio.h> // vasprintf (GNU extension)
+#include <stdio.h>  // vasprintf (GNU extension)
 #undef _GNU_SOURCE
 
-#include "pros/colors.h" // color macros
+#include "pros/colors.h"  // color macros
 
-typedef enum touch_event_e {
-  E_TOUCH_EVENT_RELEASE,
-  E_TOUCH_EVENT_PRESS,
-  E_TOUCH_EVENT_PRESS_AND_HOLD
-} touch_event_e_t;
+typedef enum touch_event_e { E_TOUCH_EVENT_RELEASE, E_TOUCH_EVENT_PRESS, E_TOUCH_EVENT_PRESS_AND_HOLD } touch_event_e_t;
 
-typedef void (*touch_event_cb_fn_t)(int32_t, int32_t);
+typedef void (*touch_event_cb_fn_t)(int16_t, int16_t);
 
 #ifdef __cplusplus
 extern "C" {
@@ -89,8 +85,7 @@ void display_scroll(int16_t start_line, int16_t lines);
  * \param lines
  *        The number of lines to scroll
  */
-void display_scroll_rect(int16_t x0, int16_t y0, int16_t x1, int16_t y1,
-                         int16_t lines);
+void display_scroll_rect(int16_t x0, int16_t y0, int16_t x1, int16_t y1, int16_t lines);
 
 /**
  * Copies a screen region from an off-screen buffer to the screen.
@@ -99,17 +94,14 @@ void display_scroll_rect(int16_t x0, int16_t y0, int16_t x1, int16_t y1,
  *        The (x,y) coordinates of the first corner of the rectangular region of
  *        the screen
  * \param x1, y1
- *        The (x,y) coordinates of the second corner of the	rectangular
- * region
+ *        The (x,y) coordinates of the second corner of the	rectangular region
  *        of the screen
  * \param buf
  *        Off-screen buffer containing screen data
  * \param stride
- *        Off-screen buffer width in pixels, such that image size is
- * stride-padding
+ *        Off-screen buffer width in pixels, such that image size is stride-padding
  */
-void display_copy_rect(int16_t x0, int16_t y0, int16_t x1, int16_t y1,
-                       uint32_t *buf, int32_t stride);
+void display_copy_rect(int16_t x0, int16_t y0, int16_t x1, int16_t y1, uint32_t* buf, int32_t stride);
 
 /**
  * Draws a pixel on the screen using the current foreground color.
@@ -217,7 +209,7 @@ void display_fill_circle(int16_t x, int16_t y, int16_t r);
  * \param ...
  *        Optional list of arguments for the format string
  */
-void display_printf(const int16_t line, const char *fmt, ...);
+void display_printf(const int16_t line, const char* fmt, ...);
 
 /**
  * Prints a formatted string to the screen at the specified coordinates.
@@ -229,7 +221,7 @@ void display_printf(const int16_t line, const char *fmt, ...);
  * \param ...
  *        Optional list of arguments for the format string
  */
-void display_printf_at(int16_t x, int16_t y, const char *fmt, ...);
+void display_printf_at(int16_t x, int16_t y, const char* fmt, ...);
 
 /**
  * Prints a large, formatted string to the screen on the specified line.
@@ -241,7 +233,7 @@ void display_printf_at(int16_t x, int16_t y, const char *fmt, ...);
  * \param ...
  *        Optional list of arguments for the format string
  */
-void display_big_printf(const int16_t line, const char *fmt, ...);
+void display_big_printf(const int16_t line, const char* fmt, ...);
 
 /**
  * Prints a large, formatted string to the screen at the specified coordinates.
@@ -253,7 +245,7 @@ void display_big_printf(const int16_t line, const char *fmt, ...);
  * \param ...
  *        Optional list of arguments for the format string
  */
-void display_big_printf_at(int16_t x, int16_t y, const char *fmt, ...);
+void display_big_printf_at(int16_t x, int16_t y, const char* fmt, ...);
 
 /**
  * Prints a small, formatted string to the screen at the specified coordinates.
@@ -265,7 +257,7 @@ void display_big_printf_at(int16_t x, int16_t y, const char *fmt, ...);
  * \param ...
  *        Optional list of arguments for the format string
  */
-void display_small_printf_at(int16_t x, int16_t y, const char *fmt, ...);
+void display_small_printf_at(int16_t x, int16_t y, const char* fmt, ...);
 
 /**
  * Prints a centered, formatted string to the screen on the specified line.
@@ -277,11 +269,10 @@ void display_small_printf_at(int16_t x, int16_t y, const char *fmt, ...);
  * \param ...
  *        Optional list of arguments for the format string
  */
-void display_center_printf(const int16_t line, const char *fmt, ...);
+void display_center_printf(const int16_t line, const char* fmt, ...);
 
 /**
- * Prints a large, centered, formatted string to the screen on the specified
- * line.
+ * Prints a large, centered, formatted string to the screen on the specified line.
  *
  * \param line
  *        The line number on which to print
@@ -290,7 +281,7 @@ void display_center_printf(const int16_t line, const char *fmt, ...);
  * \param ...
  *        Optional list of arguments for the format string
  */
-void display_center_big_printf(const int16_t line, const char *fmt, ...);
+void display_center_big_printf(const int16_t line, const char* fmt, ...);
 
 /**
  * Prints a string to the screen on the specified line.
@@ -300,7 +291,7 @@ void display_center_big_printf(const int16_t line, const char *fmt, ...);
  * \param text
  *        The text to display
  */
-void display_puts(const int16_t line, const char *text);
+void display_puts(const int16_t line, const char* text);
 
 /**
  * Prints a string to the screen at the specified coordinates.
@@ -310,7 +301,7 @@ void display_puts(const int16_t line, const char *text);
  * \param text
  *        The text to display
  */
-void display_puts_at(int16_t x, int16_t y, const char *text);
+void display_puts_at(int16_t x, int16_t y, const char* text);
 
 /**
  * Prints a large string to the screen on the specified line.
@@ -320,7 +311,7 @@ void display_puts_at(int16_t x, int16_t y, const char *text);
  * \param text
  *        The text to display
  */
-void display_big_puts(const int16_t line, const char *text);
+void display_big_puts(const int16_t line, const char* text);
 
 /**
  * Prints a large string to the screen at the specified coordinates.
@@ -330,7 +321,7 @@ void display_big_puts(const int16_t line, const char *text);
  * \param text
  *        The text to display
  */
-void display_big_puts_at(int16_t x, int16_t y, const char *text);
+void display_big_puts_at(int16_t x, int16_t y, const char* text);
 
 /**
  * Prints a small string to the screen at the specified coordinates.
@@ -340,7 +331,7 @@ void display_big_puts_at(int16_t x, int16_t y, const char *text);
  * \param text
  *        The text to display
  */
-void display_small_puts_at(int16_t x, int16_t y, const char *text);
+void display_small_puts_at(int16_t x, int16_t y, const char* text);
 
 /**
  * Prints a centered string on the specified line.
@@ -350,7 +341,7 @@ void display_small_puts_at(int16_t x, int16_t y, const char *text);
  * \param text
  *        The text to display
  */
-void display_center_puts(const int16_t line, const char *text);
+void display_center_puts(const int16_t line, const char* text);
 
 /**
  * Prints a large, centered string on the specified line.
@@ -360,7 +351,7 @@ void display_center_puts(const int16_t line, const char *text);
  * \param text
  *        The text to display
  */
-void display_center_big_puts(const int16_t line, const char *text);
+void display_center_big_puts(const int16_t line, const char* text);
 
 /**
  * Prints a formatted string to the screen on the specified line.
@@ -375,7 +366,7 @@ void display_center_big_puts(const int16_t line, const char *text);
  * \param args
  *        List of arguments for the format string
  */
-void display_vprintf(const int16_t line, const char *fmt, va_list args);
+void display_vprintf(const int16_t line, const char* fmt, va_list args);
 
 /**
  * Prints a formatted string to the screen at the specified coordinates.
@@ -390,7 +381,7 @@ void display_vprintf(const int16_t line, const char *fmt, va_list args);
  * \param args
  *        List of arguments for the format string
  */
-void display_vprintf_at(int16_t x, int16_t y, const char *fmt, va_list args);
+void display_vprintf_at(int16_t x, int16_t y, const char* fmt, va_list args);
 
 /**
  * Prints a large, formatted string to the screen on the specified line.
@@ -405,7 +396,7 @@ void display_vprintf_at(int16_t x, int16_t y, const char *fmt, va_list args);
  * \param args
  *        List of arguments for the format string
  */
-void display_big_vprintf(const int16_t line, const char *fmt, va_list args);
+void display_big_vprintf(const int16_t line, const char* fmt, va_list args);
 
 /**
  * Prints a large, formatted string to the screen at the specified coordinates.
@@ -420,14 +411,12 @@ void display_big_vprintf(const int16_t line, const char *fmt, va_list args);
  * \param args
  *        List of arguments for the format string
  */
-void display_big_vprintf_at(int16_t x, int16_t y, const char *fmt,
-                            va_list args);
+void display_big_vprintf_at(int16_t x, int16_t y, const char* fmt, va_list args);
 
 /**
  * Prints a small, formatted string to the screen at the specified coordinates.
  *
- * Same as display_small_printf_at except that this uses a va_list instead of
- * the
+ * Same as display_small_printf_at except that this uses a va_list instead of the
  * ellipsis operator so this can be used by other functions.
  *
  * \param x, y
@@ -437,8 +426,7 @@ void display_big_vprintf_at(int16_t x, int16_t y, const char *fmt,
  * \param args
  *        List of arguments for the format string
  */
-void display_small_vprintf_at(int16_t x, int16_t y, const char *fmt,
-                              va_list args);
+void display_small_vprintf_at(int16_t x, int16_t y, const char* fmt, va_list args);
 
 /**
  * Prints a centered, formatted string to the screen on the specified line.
@@ -453,11 +441,10 @@ void display_small_vprintf_at(int16_t x, int16_t y, const char *fmt,
  * \param args
  *        List of arguments for the format string
  */
-void display_center_vprintf(const int16_t line, const char *fmt, va_list args);
+void display_center_vprintf(const int16_t line, const char* fmt, va_list args);
 
 /**
- * Prints a large, centered, formatted string to the screen on the specified
- * line.
+ * Prints a large, centered, formatted string to the screen on the specified line.
  *
  * Same as display_printf except that this uses a va_list instead of the
  * ellipsis operator so this can be used by other functions.
@@ -469,8 +456,7 @@ void display_center_vprintf(const int16_t line, const char *fmt, va_list args);
  * \param args
  *        List of arguments for the format string
  */
-void display_center_big_vprintf(const int16_t line, const char *fmt,
-                                va_list args);
+void display_center_big_vprintf(const int16_t line, const char* fmt, va_list args);
 
 /**
  * Registers a user-defined callback function for an event type.
@@ -479,34 +465,29 @@ void display_center_big_vprintf(const int16_t line, const char *fmt,
  * be invoked.
  *
  * \param cb
- *        A callback function of type touch_event_cb_fn_t (void (*cb)(int16_t,
- * int16_t))
+ *        A callback function of type touch_event_cb_fn_t (void (*cb)(int16_t, int16_t))
  * \param event_type
  *        The desired type of event to listen for
  */
-void register_touch_callback(touch_event_cb_fn_t cb,
-                             touch_event_e_t event_type);
+void register_touch_callback(touch_event_cb_fn_t cb, touch_event_e_t event_type);
 
 /**
  * Unregisters a user-defined callback for an event type.
  *
- * Subsequent events of the specified type will no longer invoke the
- * user-provided
+ * Subsequent events of the specified type will no longer invoke the user-provided
  * callback when fired.
  *
  * \param cb
- *        A callback function of type touch_event_cb_fn_t (void (*cb)(int16_t,
- * int16_t))
+ *        A callback function of type touch_event_cb_fn_t (void (*cb)(int16_t, int16_t))
  * \param event_type
  *        The event type associated with the callback
  */
-void unregister_touch_callback(touch_event_cb_fn_t cb,
-                               touch_event_e_t event_type);
+void unregister_touch_callback(touch_event_cb_fn_t cb, touch_event_e_t event_type);
 
 #ifdef __cplusplus
 }
 }
 }
-#endif // extern "C"
+#endif  // extern "C"
 
 #endif
